Fix date parsing
Rather than relying on js/Date.parse which is not cross platform, just pull out
the pieces and create a new Date instance ourselves. This also means we don't
have to go back and forth to UNIX timestamps, which is convenient.

This should also address some time zone issues, albeit in a hacky
way. (js/Date.) will create a Date object in the local time zone, converting
that to UTC under the hood. Our incoming times are in CEST, so if a person is
not in CEST that will technically yield an incorrect date. *BUT* date method
like getHours will in turn convert back to the local time zone, so we get back
what we put in, which is what we want.